azure_project/function_app/requirements.txt
azure-functions
pyodbc==4.0.34
python-jose[cryptography] 